+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 
 - ``route`` and ``command`` may be used directly, not only as decorator
 - add support class-based views for ``route``
+- add ``get_app_label`` shortcut
 
 1.0.0 - 2016-09-30
 ==================